make clippy happy

+++ nativeshell/src/shell/platform/win32/drag_com.rs
@@ -474,7 +474,7 @@ impl DropSource {
     ) -> ::windows::core::HRESULT {
         if f_escape_pressed.as_bool() {
             DRAGDROP_S_CANCEL
-        } else if grf_key_state & MK_LBUTTON as u32 == 0 {
+        } else if grf_key_state & MK_LBUTTON == 0 {
             DRAGDROP_S_DROP
         } else {
             S_OK
